bug fix - cached pokeball count Updated using a svcs for every capture call.
